% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_tex.R
\name{add_numprint}
\alias{add_numprint}
\title{Add numprint to appropriate numbers}
\usage{
add_numprint(tex)
}
\arguments{
\item{tex}{The tex character vector}
}
\value{
tex character vector
}
\description{
This function
}
