% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_tex.R
\name{apply_patch.chapter}
\alias{apply_patch.chapter}
\title{Apply_patch}
\usage{
\method{apply_patch}{chapter}(x)
}
\arguments{
\item{x}{A chapter}
}
\value{
the path to typeset.tex
}
\description{
Apply the patch to the chapter
}
