% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_tex.R
\name{apply_patch}
\alias{apply_patch}
\title{Apply patch}
\usage{
apply_patch(verbose = TRUE, force = FALSE)
}
\arguments{
\item{verbose}{give information about the process.}

\item{force}{return NULL from function even if patch doesn't apply cleanly}
}
\value{
invisible(NULL)
}
\description{
Apply patch
}
