% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{assets}
\alias{assets}
\title{Get Assets}
\usage{
assets(x)
}
\arguments{
\item{x}{The report object, chapter object or filename.}
}
\value{
character string with path to assets.
}
\description{
Determine the assets directory given a report, chapter or chapter
file in the report project.
}
