% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{asterisk}
\alias{asterisk}
\title{Add asterisk to sections}
\usage{
asterisk(tex, direction = c("add", "remove"))
}
\arguments{
\item{tex}{The tex character vector}
}
\value{
tex character vector
}
\description{
Add asterisk to sections
}
\keyword{internal}
