% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter_text.R
\name{build_text}
\alias{build_text}
\title{Build chapters}
\usage{
build_text(x)
}
\arguments{
\item{x}{The report object, chapter object or the path to the text
tex file.}
}
\value{
invisible NULL
}
\description{
Build chapters
}
