% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{c_tex}
\alias{c_tex}
\title{Concatenate .tex files}
\usage{
c_tex(tex)
}
\arguments{
\item{tex}{A vector of .tex file paths}
}
\value{
The filename to the temporary file with the embeded tex.
}
\description{
Concatenate .tex files
}
\keyword{internal}
