% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{chapter_tex_files}
\alias{chapter_tex_files}
\title{Get the chapter tex files}
\usage{
chapter_tex_files(type = c("all", "text", "fig", "table"))
}
\description{
Get the chapter tex files
}
\keyword{internal}
