% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Check report}
\usage{
check(path = ".")
}
\arguments{
\item{path}{The path to the report.}
}
\value{
invisible \code{FALSE} if OK, else invisible \code{TRUE}.
}
\description{
Check report
}
