% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_apply_typeset_patch}
\alias{check_apply_typeset_patch}
\title{Check apply patch 'typeset.patch' to 'text.tex'}
\usage{
check_apply_typeset_patch()
}
\description{
Checking that applying patches doesn't generate warnings or
errors.
}
\keyword{internal}
