% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_expect_clean_repository}
\alias{check_expect_clean_repository}
\title{Check that repository is clean}
\usage{
check_expect_clean_repository(x)
}
\description{
To protect against overwriting un-committed changes.
}
\keyword{internal}
