% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_expect_pandoc_is_installed}
\alias{check_expect_pandoc_is_installed}
\title{Check that pandoc is installed}
\usage{
check_expect_pandoc_is_installed()
}
\description{
pandoc is required to check for example 'tex' to 'docx' round
trip.
}
\keyword{internal}
