% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_expect_patch_is_installed}
\alias{check_expect_patch_is_installed}
\title{Check that GNU patch is installed}
\usage{
check_expect_patch_is_installed()
}
\description{
patch is required to check that applying patches work.
}
\keyword{internal}
