% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_expect_readxl_is_installed}
\alias{check_expect_readxl_is_installed}
\title{Check that readxl is installed}
\usage{
check_expect_readxl_is_installed()
}
\description{
readxl is required to check consistency between 'report.yml' and
exportd 'Authors.xlsx'.
}
\keyword{internal}
