% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_missing_table_reference_files}
\alias{check_missing_table_reference_files}
\title{Check for missing table reference files}
\usage{
check_missing_table_reference_files(x)
}
\arguments{
\item{x}{the report or chapter object.}
}
\description{
Check for table references in the 'text.tex' file that does not
have a corresponding 'table.tex' file.
}
\keyword{internal}
