% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_reference_format}
\alias{check_reference_format}
\title{Check reference format}
\usage{
check_reference_format(x)
}
\arguments{
\item{x}{the report or chapter object.}
}
\description{
Check reference format
}
\keyword{internal}
