% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_tex_to_docx_round_trip}
\alias{check_tex_to_docx_round_trip}
\title{Check conversion between 'tex' and 'docx'}
\usage{
check_tex_to_docx_round_trip(x)
}
\description{
Checking that converting to 'docx' from 'tex' and converting back
to 'tex' doesn't generate changes.
}
\keyword{internal}
