% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{clean_tex}
\alias{clean_tex}
\title{Clean a chapter 'text.tex' file}
\usage{
clean_tex(path)
}
\arguments{
\item{path}{The path to the chapter.}
}
\value{
The filename to the temporary file.
}
\description{
Clean a chapter 'text.tex' file and save result to temporary file.
}
\keyword{internal}
