% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrTeX.R
\name{col_widths}
\alias{col_widths}
\title{Column widths}
\usage{
col_widths(tbl)
}
\arguments{
\item{tbl}{table to get the column widths from.}
}
\value{
numeric vector with column widths.
}
\description{
Get the columns widths of a docx table.
}
