% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relax.R
\name{contributors}
\alias{contributors}
\title{Contributors}
\usage{
contributors(sheet)
}
\arguments{
\item{sheet}{a data.frame defining the project.}
}
\value{
a list with contributors
}
\description{
Extract the contributors from the project excel sheet.
}
