% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{convert_docx_ref_to_href}
\alias{convert_docx_ref_to_href}
\title{Convert the docx references to tex href}
\usage{
convert_docx_ref_to_href(tex, title)
}
\arguments{
\item{tex}{The tex character vector}

\item{title}{The chapter title}
}
\value{
tex character vector
}
\description{
Convert the docx references to tex href
}
\keyword{internal}
