% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{convert_ref_to_docx_ref}
\alias{convert_ref_to_docx_ref}
\title{Convert the tex ref to docx ref}
\usage{
convert_ref_to_docx_ref(tex)
}
\arguments{
\item{tex}{The tex character vector}
}
\value{
tex character vector
}
\description{
Convert the tex ref to docx ref
}
\keyword{internal}
