% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Makefile.R
\name{create_Makefile}
\alias{create_Makefile}
\title{Create a Makefile in each chapter of the report}
\usage{
create_Makefile(x, ...)
}
\arguments{
\item{x}{The report object}

\item{...}{Additional arguments.}

\item{repo}{The git repository to add the 'Makefile' to.}
}
\value{
invisible NULL
}
\description{
Create a Makefile in each chapter of the report
}
