% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrTeX.R
\name{dim.docx_w_tc}
\alias{dim.docx_w_tc}
\title{Dimensions of a cell.}
\usage{
\method{dim}{docx_w_tc}(x)
}
\arguments{
\item{x}{the cell to get the dimension from}
}
\description{
Note that if the cell is part of a vertically merged region, then
the first dimension is \code{NA}.
}
