% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{do_init}
\alias{do_init}
\title{Initialize a report structure}
\usage{
do_init(x, repo, import)
}
\arguments{
\item{x}{The report object to initialize.}

\item{repo}{The git repository to add the report to.}

\item{import}{The root path to another report to import chapters
from.}
}
\value{
invisible NULL.
}
\description{
Initialize a report structure
}
\keyword{internal}
