% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrTeX.R
\name{docx_caption}
\alias{docx_caption}
\title{Create a docx caption object}
\usage{
docx_caption(xml)
}
\arguments{
\item{xml}{the xml node for the docx caption.}
}
\value{
a \code{docx_caption} object.
}
\description{
Create a docx caption object
}
