% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrTeX.R
\name{docx_footnote}
\alias{docx_footnote}
\title{Create a docx footnote object}
\usage{
docx_footnote(xml)
}
\arguments{
\item{xml}{the xml node for the docx footnotes.}
}
\value{
a \code{docx_footnote} object.
}
\description{
Create a docx footnote object
}
