% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrTeX.R
\name{docx_label}
\alias{docx_label}
\title{Create a docx label object}
\usage{
docx_label(xml)
}
\arguments{
\item{xml}{the xml node for the docx label.}
}
\value{
a \code{docx_label} object.
}
\description{
Create a docx label object
}
