% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrTeX.R
\name{docx_paragraph}
\alias{docx_paragraph}
\title{Create a docx paragraph object}
\usage{
docx_paragraph(xml)
}
\arguments{
\item{xml}{the xml node for the docx paragraph.}
}
\value{
a \code{docx_paragraph} object.
}
\description{
Create a docx paragraph object
}
