% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrTeX.R
\name{docx_table}
\alias{docx_table}
\title{Create a docx table object}
\usage{
docx_table(xml)
}
\arguments{
\item{xml}{the xml node for the docx table.}
}
\value{
a \code{docx_table} object.
}
\description{
A \code{docx_table} consists of a caption, a label, table data and
footnotes.
}
