% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrTeX.R
\name{docx_tables}
\alias{docx_tables}
\title{Load all tables in a docx file}
\usage{
docx_tables(filename)
}
\arguments{
\item{filename}{name of the docx file.}
}
\description{
Load all tables in a docx file
}
\examples{
## Load tables from a bundled docx file.
tbl <- docx_tables(system.file("extdata/table.docx", package = "doctex"))

## View the layout of the tables
tbl

## View the column widths
sapply(tbl, col_widths)
}
