% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{embed_tex}
\alias{embed_tex}
\title{Embed a tex file in a context}
\usage{
embed_tex(tex, pre_tex, post_tex)
}
\arguments{
\item{tex}{The tex file to embed.}

\item{pre_tex}{A tex snippet to add before the tex file.}

\item{post_tex}{A tex snippet to add after the tex file.}
}
\value{
The filename to the temporary file with the embeded tex.
}
\description{
Embed a tex file in a context
}
\keyword{internal}
