% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{export}
\alias{export}
\title{Export files to workspace}
\usage{
export(x)
}
\arguments{
\item{x}{The object to export. The docx files are exported to
workspace/chapters/title.}
}
\value{
invisible NULL.
}
\description{
Authors make contributions to chapters in the workspace.
}
