% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{figure_files}
\alias{figure_files}
\title{List figure files}
\usage{
figure_files(x, fileext = "all")
}
\arguments{
\item{fileext}{The extension of the files you want to find}
}
\description{
A method to discover the figures in a report or chapter object.
}
