% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{format_labels}
\alias{format_labels}
\title{Format the label from a figure path}
\usage{
format_labels(labels, format = c("latex", "word"))
}
\description{
Format the label from a figure path
}
\keyword{internal}
