% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mill.R
\name{formatted_authors}
\alias{formatted_authors}
\title{Formatting authors}
\usage{
formatted_authors(x, format = c("name", "email"))
}
\arguments{
\item{x}{A report object}

\item{format}{The format of the author list you want}
}
\value{
A unique sorted list of authors
}
\description{
This function gets a unique list of authors and sorts them
alphabetically. I assume this should actually be solved as a print
method and format method for an authors class object.
}
\details{
We also have a lexigraphical sort problem in this function that is
solved with a hack that might have a better cross platform
solution.
}
\author{
Thomas Rosendal
}
