% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{from_docx}
\alias{from_docx}
\title{Convert from docx to tex}
\usage{
from_docx(repo = NULL, force = FALSE)
}
\arguments{
\item{repo}{the report git repository.}

\item{force}{if \code{TRUE}, run even if working tree is not
clean. Default is \code{FALSE}.}
}
\value{
invisible NULL.
}
\description{
Use pandoc (http://pandoc.org/) to convert from 'docx' to
'tex'. The chapter 'text.docx' is converted to 'text.tex'. Each
chapter 'text.tex' is added, but not commited, to the report git
repository.
}
