% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{get_labels}
\alias{get_labels}
\title{Get figure labels}
\usage{
get_labels(x, what, reftype, include)
}
\arguments{
\item{x}{The report object or chapter object}
}
\value{
invisible NULL
}
\description{
Get figure labels
}
