% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{image_diff}
\alias{image_diff}
\title{Compare Images}
\usage{
image_diff(reference, new, dir = tempdir())
}
\arguments{
\item{reference}{Path to a reference image}

\item{new}{Path to an image to compare to the reference image}

\item{dir}{directory to perform the comparison inside}
}
\value{
numeric The percent difference between the two images
}
\description{
Submit an a reference image and a new image and get back the
percent difference in all colour channels.
}
\details{
Requires imagemagick
}
