% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{import}
\alias{import}
\title{Import files}
\usage{
import(x, from)
}
\arguments{
\item{x}{The object to import.}

\item{from}{The source of the import. If the argument is missing,
the docx files are imported from a folder named from the
report title.}
}
\value{
invisible NULL.
}
\description{
Import files
}
