% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{import_from}
\alias{import_from}
\title{Import a chapter from another report}
\usage{
import_from(from, to, title, filename)
}
\arguments{
\item{from}{The path to the root folder of the other report to
import the chapter from.}

\item{to}{The path of the chapter to import to.}

\item{title}{The title of the chapter.}

\item{filename}{The filename of the file to import.}
}
\value{
TRUE if the file to import existed, else FALSE.
}
\description{
Import a chapter from another report
}
\keyword{internal}

