% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_clean}
\alias{init_clean}
\title{Clean folder structure before init}
\usage{
init_clean(path, force)
}
\arguments{
\item{path}{The path to the folder to remove.}

\item{force}{The method fails if the folder exists and force is
not TRUE.}
}
\value{
invisible NULL
}
\description{
Clean folder structure before init
}
\keyword{internal}
