% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_tex.R
\name{interactive_replace.character}
\alias{interactive_replace.character}
\title{Interactive replacement}
\usage{
\method{interactive_replace}{character}(x, pattern, replacement)
}
\arguments{
\item{x}{string}

\item{pattern}{A regular expression with a single capture}

\item{replacement}{The replacement expression Referring to the
first capture with "\\1".}
}
\value{
string
}
\description{
Takes a character vector of length 1 and searches for the pattern
and replaces it with the replacement.
}
\keyword{internal}
