% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_tex.R
\name{interactive_replace}
\alias{interactive_replace}
\title{Interactive find and replace over a chapter}
\usage{
interactive_replace(x, pattern, replacement)
}
\arguments{
\item{x}{The tex character vector}

\item{pattern}{A regular expression with a single capture}

\item{replacement}{The replacement expression Referring to the
first capture with "\\1".}
}
\value{
tex character vector
}
\description{
This function does the interactive replacement for a chapter. An
example to add numprint{} might be:
}
\examples{
r <- load_report()
pattern <- "([[:digit:]]{4, })"
replacement <- " \\\\\\numprint{\\\\1}"
interactive_replace(r[["Campy"]], pattern, replacement) 
}
