% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relax.R
\name{load_report}
\alias{load_report}
\title{Load configuration for the report}
\usage{
load_report(path = ".")
}
\arguments{
\item{path}{The path to the root folder of the project.}
}
\description{
Load configuration for the report
}

