% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{luatex}
\alias{luatex}
\title{Run LuaTeX}
\usage{
luatex(texname, clean = TRUE)
}
\arguments{
\item{texname}{Run LuaTeX on texname.}

\item{clean}{logical. If \code{TRUE}, auxiliary files (aux, log,
out) created by LuaTeX are removed.}
}
\value{
invisible NULL.
}
\description{
Run LuaTeX
}
\keyword{internal}
