% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{luatex_cmd}
\alias{luatex_cmd}
\title{Select a luatex system call}
\usage{
luatex_cmd(options = c("--interaction=nonstopmode", "-V 7"))
}
\value{
character string, giving the luatex command to run. On
    Windows: \code{texify --pdf --engine=luatex
    --max-iterations=50}, else \code{lualatex}.
}
\description{
Select a luatex system call
}
\keyword{internal}
