% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{make_labels_chapter_specific}
\alias{make_labels_chapter_specific}
\title{Make tex labels chapter specific in the report}
\usage{
make_labels_chapter_specific(tex, title)
}
\arguments{
\item{tex}{The tex character vector}

\item{title}{The chapter title}
}
\value{
tex character vector
}
\description{
Make tex labels chapter specific in the report
}
\keyword{internal}
