% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mill.R
\docType{data}
\name{NUTS_10M}
\alias{NUTS_10M}
\title{Administrative (NUTS) units in Sweden (1 : 10 Million)}
\source{
\url{http://epp.eurostat.ec.europa.eu/portal/page/portal/gisco_Geographical
_information_maps/popups/references/administrative_units
_statistical_units_1}
}
\description{
The dataset contains administrative units (NUTS) in Sweden in
European terminology in scale 1 : 10 Million.
}
\keyword{datasets}
