% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{pdf_diff}
\alias{pdf_diff}
\title{pdf_diff}
\usage{
pdf_diff(reference, new, dir = tempdir())
}
\arguments{
\item{reference}{Path to a pdf file}

\item{new}{Path to a pdf file}

\item{dir}{directory to perform the comparison inside}
}
\value{
data.frame A data.frame with 3 columns: page,
    percent_diff, composite.
}
\description{
Difference of two pdf:s
}
\details{
Submitt two pdfs of the same number of pages to this function and
get back a data.frame with the same number of rows as the pdf with
the percent pixel difference for each page and a path to a
composed image of each page highlighting the differences in red.
}
