% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{pdf_np}
\alias{pdf_np}
\title{pdf_np}
\usage{
pdf_np(path)
}
\arguments{
\item{path}{the path to the pdf file}
}
\value{
numeric The number of pages
}
\description{
Number of pages in a pdf
}
\details{
Get the number of pages of a pdffile

Requires pdftk
}
