% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{plot.pdf_diff_df}
\alias{plot.pdf_diff_df}
\title{plot.pdf_diff_df

A plotting method for the dataframe generated by 'pdf_diff()'}
\usage{
\method{plot}{pdf_diff_df}(x, ...)
}
\arguments{
\item{x}{the data.frame generated by pdf_diff()}

\item{...}{other arguments to the genertic plot function}
}
\value{
A plot
}
\description{
plot.pdf_diff_df

A plotting method for the dataframe generated by 'pdf_diff()'
}
