% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{preview_figures}
\alias{preview_figures}
\title{Preview figures}
\usage{
preview_figures(x)
}
\arguments{
\item{x}{The report object or chapter object}
}
\value{
invisible NULL
}
\description{
Preview figures
}
