% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter_text.R
\name{preview_text.chapter}
\alias{preview_text.chapter}
\title{Preview a chapter text}
\usage{
\method{preview_text}{chapter}(chapter)
}
\arguments{
\item{chapter}{The path to the text tex file}
}
\description{
Preview a chapter text
}
