% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter_text.R
\name{preview_text}
\alias{preview_text}
\title{Preview chapters}
\usage{
preview_text(x)
}
\arguments{
\item{x}{The report object, chapter object or the path to the text
tex file.}
}
\value{
invisible NULL
}
\description{
Preview chapters
}
