% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{reduce_image}
\alias{reduce_image}
\title{Reduce Images}
\usage{
reduce_image(from, to)
}
\arguments{
\item{from}{filename of the original image}

\item{to}{filename of the reduced image}
}
\value{
invisible NULL
}
\description{
Reduce the resolution of images. Requires imagemagick.
}
