% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{reference_pattern}
\alias{reference_pattern}
\title{Create a regular expression pattern to search for labels or
references}
\usage{
reference_pattern(cmd = c("label", "ref"))
}
\arguments{
\item{cmd}{what to search for, either \\label{.} or \\ref{.}.}
}
\description{
Create a regular expression pattern to search for labels or
references
}
\keyword{internal}
