% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{references}
\alias{references}
\title{TeX references}
\usage{
references(x, reftype)
}
\arguments{
\item{x}{the report or chapter object.}

\item{reftype}{the type of references to search for.}
}
\value{
character vector with found references.
}
\description{
TeX references
}
\keyword{internal}
