% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{roundtrip}
\alias{roundtrip}
\title{Roundtrip tex to docx}
\usage{
roundtrip(x, ...)
}
\arguments{
\item{x}{The object to convert.}

\item{...}{Additional arguments.}

\item{repo}{The git repository to add the 'docx' to.}
}
\value{
invisible NULL.
}
\description{
Roundtrip tex to docx
}
