% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work_status.R
\name{sparkline}
\alias{sparkline}
\title{sparkline}
\usage{
sparkline(path = ".", ylim = NULL, xlim = NULL)
}
\arguments{
\item{path}{path to repo}

\item{ylim}{ylim of graphs}

\item{xlim}{xlim of graphs}
}
\value{
plots
}
\description{
sparkline
}
