% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{step_section}
\alias{step_section}
\title{Step up the section levels}
\usage{
step_section(tex, direction = c("up", "down"))
}
\arguments{
\item{tex}{The tex character vector}

\item{direction}{go 'up' or 'down'}
}
\value{
tex character vector
}
\description{
Going 'up':
1st section -> chapter
2nd subsection -> section
3rd subsubsection -> subsection
4th paragraph -> subsubsection
}
\details{
Going 'down':
1st subsubsection -> paragraph
2nd subsection -> subsubsection
3rd section -> subsection
4th chapter -> section
}
\keyword{internal}
