% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter_text.R
\name{text_files}
\alias{text_files}
\title{Get the path(s) of the chapter text.tex file(s)}
\usage{
text_files(x)
}
\arguments{
\item{x}{The chapter object}
}
\value{
path
}
\description{
Get the path(s) of the chapter text.tex file(s)
}
\keyword{internal}
