% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_tex.R
\name{apply_patch}
\alias{apply_patch}
\title{Apply patch}
\usage{
apply_patch(verbose = TRUE)
}
\arguments{
\item{verbose}{give information about the process.}
}
\value{
invisible(NULL)
}
\description{
Apply patch
}
