% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{assets}
\alias{assets}
\title{Get the assets directory}
\usage{
assets(filename)
}
\arguments{
\item{filename}{The filename.}
}
\value{
path to the assets directory.
}
\description{
Determine the assets directory given a chapter file in the report
project.
}
\keyword{internal}
