% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relax.R
\name{contributor}
\alias{contributor}
\title{Contributor}
\usage{
contributor(row, title)
}
\arguments{
\item{row}{the row with the contributor.}

\item{title}{the title of the chapter.}
}
\value{
a contributor object
}
\description{
Extract one contributor from one row of the project excel sheet.
}
\keyword{internal}
