% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Makefile.R
\name{create_Makefile}
\alias{create_Makefile}
\title{Create a Makefile in each chapter of the report}
\usage{
create_Makefile()
}
\value{
invisible NULL
}
\description{
Create a Makefile in each chapter of the report
}
