% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{export}
\alias{export}
\title{Export files}
\usage{
export(x, to)
}
\arguments{
\item{x}{The object to export.}

\item{to}{The destination of the export. If the argument is
missing, the docx files are exported to a folder named from
the report title.}
}
\value{
invisible NULL.
}
\description{
Export files
}
