% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{from_docx}
\alias{from_docx}
\title{Convert from docx to tex}
\usage{
from_docx(x, ...)
}
\arguments{
\item{x}{The object to convert.}

\item{...}{Additional arguments.}

\item{repo}{The git repository to add the 'tex' to.}
}
\value{
invisible NULL.
}
\description{
Use pandoc (http://pandoc.org/) to convert from 'docx' to
'tex'. The chapter 'text.docx' is converted to 'text.tex'. Each
chapter 'text.tex' is added, but not commited, to the report git
repository.
}
