% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{image_diff}
\alias{image_diff}
\title{image_diff}
\usage{
image_diff(reference, new, output = tempfile())
}
\arguments{
\item{reference}{Path to a reference image}

\item{new}{Path to an image to compare to the reference image}

\item{output}{Optional path to the output image}
}
\value{
numeric The percent difference between the two images
}
\description{
Compare Images
}
\details{
Submit an a reference image and a new image and get back the
percent difference in all colour channels. Optionally you can also
save the output comparison image that highlights the differences
between them.

Requires imagemagick
}
