% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_report}
\alias{init_report}
\title{Intialize the report folder structure}
\usage{
init_report(path = ".", force = FALSE)
}
\arguments{
\item{path}{The path to the root folder of the project.}

\item{force}{The method fails if the folder structure already
exists and force equals FALSE.}
}
\description{
Intialize the report folder structure
}

