% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{pdf_split}
\alias{pdf_split}
\title{pdf_split}
\usage{
pdf_split(path, dir)
}
\arguments{
\item{path}{The path to the pdffile}

\item{dir}{directory to perform the comparison inside}
}
\value{
list A list of the filenames of the resultant 1 page pdfs
}
\description{
Split up a pdf
}
\details{
Split a pdf into it's component pages and return a list of
filenames.
}
