% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{pdf_split}
\alias{pdf_split}
\title{Split up a pdf}
\usage{
pdf_split(path, dir = tempdir())
}
\arguments{
\item{path}{The path to the pdf file to split.}

\item{dir}{directory for output. Default is to use a temporary
directory.}
}
\value{
A vector of the filenames of the resultant 1 page pdfs.
}
\description{
Split a pdf into it's component pages and return a vector of
filenames.
}
